refactor: 改名，更直观; 微调脚本功能
