+++ src/policy/pkg_cfg.rs
@@ -2,7 +2,7 @@ use super::name_match::{policy_ue, policy_unity};
 use crate::{
     activity::ActivityUtils,
     cpu_common::Controller,
-    policy::usage::{usage_top1::policy_usage1, usage_top2::policy_usage2},
+    policy::usage::{usage_top1::policy_ue_top1, usage_top2::policy_usage2},
 };
 use libc::pid_t;
 use once_cell::sync::Lazy;
@@ -16,7 +16,7 @@ const UNITY: [&str; 6] = [
     "com.tencent.tmgp.speedmobile",
 ];
 
-const UNNAME1: [&str; 3] = [
+const UE_USAGE_T1: [&str; 3] = [
     "com.tencent.lzhx",
     "com.tencent.tmgp.pubgmhd",
     "com.tencent.tmgp.pubgmhd",
@@ -44,7 +44,7 @@ type ConfigTuple<'a> = (&'a [&'a str], fn(&mut StartArgs));
 
 pub static PACKAGE_CONFIGS: Lazy<[ConfigTuple; 4]> = Lazy::new(|| {
     [
-        (&UNNAME1[..], policy_usage1::start_task),
+        (&UE_USAGE_T1[..], policy_ue_top1::start_task),
         (&UNNAME2[..], policy_usage2::start_task),
         (&UNITY[..], policy_unity::start_task),
         (&UE[..], policy_ue::start_task),
