+++ test_build.sh
@@ -1,0 +1,4 @@
+cargo fmt
+rm -rf output
+rm -rf $(find ./target -name "*thread-opt*")
+python3 ./make.py build --debug --nightly
